requirements.txt
Flask~=2.2.2
plac~=1.3.5
pandas~=2.0.1
langchain~=0.0.167
python-dotenv~=1.0.0
tqdm~=4.65.0
chromadb>=0.3.0
sentence_transformers
torch


